Time: 116 ms (97.13%), Space: 71.4 MB (62.06%) - LeetHub

+++ 2439-minimize-maximum-of-array/2439-minimize-maximum-of-array.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int minimizeArrayValue(vector<int>& nums) {
-        long long sum = 0, res = 0;
+        long sum = 0, res = 0;
         for (int i = 0; i < nums.size(); ++i) {
             sum += nums[i];
-            res = max(res, (long long)ceil((double)sum/(i+1)));
+            res = max(res, (long)ceil((double)sum/(i+1)));
         }
         return res;
     }
